refactor(core): move protocol state machine into openvtc-core (R17)
Move the pure protocol logic out of the TUI's message_dispatch.rs into a
new openvtc-core::messaging module so the state machine is testable without
the TUI crate (T1 enabler). The async I/O orchestrator
(process_inbound_message) stays in the TUI and now imports the moved
functions from core.

Moved verbatim into openvtc-core/src/messaging.rs:
- SeenMessages replay guard, unix_now, check_message_age, check_task_capacity
- handle_join_submit_receipt, handle_credential_issue
- vet_vrc_issued, verify_vrc_proof
- require_thid, validate_did, is_did_msi_char, create_finalize_message
- build_didcomm_message + MESSAGE_EXPIRY_SECS (pure; re-exported from the
  TUI's didcomm.rs so existing call sites are unchanged)
- the full #[cfg(test)] module (28 tests) — they construct only core types
  and now run under `cargo test -p openvtc-core`

Behavior identical: relocation, not a logic change.
